[#13] feat: Router & ChartPage 생성 및 App 내부의 로직 ChartPage로 이전
