+++ src/pages/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as ChartPage } from './ChartPage';
